Переделал switch на if, создал новый enum для string значений, переименовал MinionEnum в MinionIntEnum, удалил string из MinionIntEnum. Замениел string на enum в Castle и Horde.
